+++ NEWS.md
@@ -6,6 +6,7 @@
 * Added option to include covariates in sem_dadas
 * Added variance test with sem in sem_dadas
 * Added variance test via parametric bootstrap in ml_dadas
+* Added cvv_manual -function for calculation of coefficients of variance variation from manually inputted sample sizes and variances of multiple variables
 
 # multid 0.4.0